Fix indentation in class ElementTree.
Move test_api tests into test_etree.py:ETreeOnlyTestCase.
